requirements.txt
annotated-types==0.7.0
asgiref==3.8.1
diff-match-patch==20241021
dj-user-login-history==1.0.6
Django==5.1.2
django-image-uploader-widget==1.0.0
django-import-export==4.2.0
django-ninja==1.3.0
django-unfold==0.40.0
dnspython==2.7.0
email_validator==2.2.0
fastnanoid==0.4.1
idna==3.10
orjson==3.10.10
pillow==11.0.0
psycopg==3.2.3
pydantic==2.9.2
pydantic_core==2.23.4
python-decouple==3.8
sqlparse==0.5.1
tablib==3.7.0
typing_extensions==4.12.2
